+++ CHANGELOG.md
@@ -1,12 +1,31 @@
-## [1.0.4](https://github.com/MilosPaunovic/questionnaire/compare/1.0.3...1.0.4)
+# Changelog
+
+## [2.0.0](https://github.com/MilosPaunovic/questionnaire/compare/1.0.4...2.0.0) - 2022-09-30
+
+### Breaking changes
+
+This major release introduces a breaking change as the method `questions()` now expects an object parameter, instead of number, like in the previous versions.
+
+### Added
+
+* Allowed options argument to be passed for recieving wrong answers, as well as the correct one. 
+
+## [1.0.4](https://github.com/MilosPaunovic/questionnaire/compare/1.0.3...1.0.4) - 2022-09-28
+
+### Fixed
 
 * Included question data files into the final bundle.
 
-## [1.0.3](https://github.com/MilosPaunovic/questionnaire/compare/1.0.2...1.0.3)
+## [1.0.3](https://github.com/MilosPaunovic/questionnaire/compare/1.0.2...1.0.3) - 2022-09-28
 
-* Trimmed down list of questions following the [advice](https://www.linkedin.com/feed/update/urn:li:activity:6980753295447330816?commentUrn=urn%3Ali%3Acomment%3A%28activity%3A6980753295447330816%2C6980847167380803586%29) of the initial answers creator.
+### Changed
+
 * Improved the way of library bundling.
 
-## [1.0.2](https://github.com/MilosPaunovic/questionnaire/commits/1.0.2)
+### Removed
 
-* Initial release with two variations of questions in Serbian (Latin and Cyrillic) and two methods, question and questions.
+* Trimmed down list of questions following the [advice](https://www.linkedin.com/feed/update/urn:li:activity:6980753295447330816?commentUrn=urn%3Ali%3Acomment%3A%28activity%3A6980753295447330816%2C6980847167380803586%29) of the initial answers creator.
+
+## [1.0.2](https://github.com/MilosPaunovic/questionnaire/commits/1.0.2) - 2022-09-28
+
+_Initial release._